% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_eyes_correction.R
\name{big_eyes_correction}
\alias{big_eyes_correction}
\title{Correct Big Eyes bearing readings}
\usage{
big_eyes_correction(bearing)
}
\arguments{
\item{bearing}{Big Eyes bearing reading, in degrees.}
}
\value{
Corrected bearing.
}
\description{
Correct Big Eyes bearing readings
}
