% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_divider.R
\name{body_divider}
\alias{body_divider}
\title{Title}
\usage{
body_divider(
  steps,
  from,
  to,
  n,
  width = 3860,
  height = 2700,
  length_frac = 0.2,
  color = "black",
  line_width = 1,
  lty = 1
)
}
\arguments{
\item{steps}{desc}

\item{from}{desc}

\item{to}{desc}

\item{n}{desc}

\item{width}{desc}

\item{height}{desc}

\item{length_frac}{desc}

\item{color}{desc}

\item{line_width}{desc}

\item{lty}{desc}
}
\value{
desc
}
\description{
Title
}
