% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_extension.R
\name{body_extension}
\alias{body_extension}
\title{Title}
\usage{
body_extension(
  steps,
  reference,
  extension_factor = 4,
  one_sided = 0,
  color = "black",
  line_width = 1,
  lty = 1
)
}
\arguments{
\item{steps}{desc}

\item{reference}{desc}

\item{extension_factor}{desc}

\item{one_sided}{desc}

\item{color}{desc}

\item{line_width}{desc}

\item{lty}{desc}
}
\value{
desc
}
\description{
Title
}
