% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_line.R
\name{body_line}
\alias{body_line}
\title{Title}
\usage{
body_line(
  steps,
  measure_sequence = c("rostrum", "blowhole", "dorsal fin", "peduncle",
    "insertion point"),
  color = "black",
  line_width = 1
)
}
\arguments{
\item{steps}{desc}

\item{measure_sequence}{desc}

\item{color}{desc}

\item{line_width}{desc}
}
\value{
desc
}
\description{
Title
}
