% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_orthogonal.R
\name{body_orthogonal}
\alias{body_orthogonal}
\title{Title}
\usage{
body_orthogonal(
  steps,
  intersection,
  reference1 = NULL,
  reference2 = NULL,
  width = 3860,
  height = 2700,
  length_frac = 0.2,
  color = "black",
  line_width = 1,
  lty = 1
)
}
\arguments{
\item{steps}{desc}

\item{intersection}{desc}

\item{reference1}{desc}

\item{reference2}{desc}

\item{width}{desc}

\item{height}{desc}

\item{length_frac}{desc}

\item{color}{desc}

\item{line_width}{desc}

\item{lty}{desc}
}
\value{
desc
}
\description{
Title
}
