% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions_sync.R
\name{conditions_sync}
\alias{conditions_sync}
\title{Association conditions to a detection}
\usage{
conditions_sync(df, sea, verbose = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} with at least these columns: \code{scan_key} (unique ID for the scan),
\code{bearing} (in degrees), and \code{km_sit} (distance to detection in km).}

\item{sea}{A \code{data.frame} produced by \code{survey_overviews()} or \code{survey_overview()}.}

\item{verbose}{Boolean: print progress updates to Console?}
}
\value{
Expanded dataframe.
}
\description{
Association conditions to a detection
}
