% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_rates.R
\name{detection_rates}
\alias{detection_rates}
\title{Calculate weekly detection rates}
\usage{
detection_rates(scans, sits, min_vis = 50, max_bft = 3)
}
\arguments{
\item{scans}{A \code{data.frame} produced by \code{survey_overviews()} or \code{survey_overview()} and
processed with \code{effort_conditions()}.}

\item{sits}{A \code{data.frame} produced by \code{survey_overviews()} or \code{survey_overview()}.}

\item{min_vis}{Minimum acceptable percent-visibility for a scan in order to
include in a detection rate calculation.}

\item{max_bft}{Maximum acceptable Beaufort sea state for a scan in order to
include in a detection rate calculation.}
}
\value{
a \code{data.frame} with detection summaries and rates )detections per hour) for each species.
}
\description{
Calculate weekly detection rates
}
