% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.r
\name{dispersion}
\alias{dispersion}
\alias{dispersion.survey}
\title{Calculating cross-sectional measures of forecast dispersion}
\usage{
dispersion(obj, ...)

\method{dispersion}{survey}(surveyObj, method = 1)
}
\arguments{
\item{surveyObj}{an object of class survey}

\item{method}{the dispersion measure}
}
\value{
The function returns an object of class ts
}
\description{
Calculating cross-sectional measures of forecast dispersion
}
\details{
Computes dispersion measures for forecasters.
\itemize{
\item{1}{Interquartile Range}
\item{2}{Standard deviation}
\item{3}{Interquartile Range on Q/Q-growth}
}
}
\examples{
tmp <- individualSPF(survey = "CPI", variable = "CPI3")
dsp <- dispersion(tmp,method = 2)
}
