% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.r
\name{downloadSPF}
\alias{downloadSPF}
\title{Download survey data}
\usage{
downloadSPF(survey = "NGDP", type = "mean")
}
\arguments{
\item{survey}{The survey you wish to download}

\item{type}{The type you wish you to download, i.e. mean, median or annualized growth rates}
}
\value{
downloadSPF returns an object of class "survey".

An object of class "survey" is a list containing the following components
\item{survey}{The name of the survey}
\item{type}{Type of the survey, i.e. mean, median, growth, or individual (see individualSPF)}
\item{variable}{Name of the variable indicating the forecast horizon, e.g. "NGDP1", or "NGDP2", downloaded. If "all", all forecast horizons were downloaded}
\item{sseries}{An object of the class ts containing the downloaded series}
}
\description{
Download survey data
}
\examples{
tmp <- downloadSPF(survey = "CPI", type = "growth")
}
