% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_sync.R
\name{effort_sync}
\alias{effort_sync}
\title{Association scan effort to a detection}
\usage{
effort_sync(df, scans, verbose = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} with at least these columns: \code{scan_key} (unique ID for the scan),
\code{bearing} (in degrees), and \code{km_sit} (distance to detection in km).}

\item{verbose}{Boolean: print progress updates to Console?}

\item{sea}{A \code{data.frame} produced by \code{survey_overviews()} or \code{survey_overview()}.}
}
\value{
The \code{df} input with new conditions columns.
}
\description{
Association scan effort to a detection
}
