% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdistr.r
\name{fit_distribution}
\alias{fit_distribution}
\title{fit survey to distribution}
\usage{
fit_distribution(srvObj, year, quarter, id, distr = "Beta", ...)
}
\arguments{
\item{srvObj}{survey}

\item{year}{year of forecast}

\item{quarter}{quarter of forecast}

\item{id}{forecaster}

\item{distr}{probability (cu)}
}
\description{
fit survey to distribution
}
