% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_circle_distance.R
\name{great_circle_distance}
\alias{great_circle_distance}
\title{Great circle distance based on angle}
\usage{
great_circle_distance(tilt, H, R = 6364.5)
}
\arguments{
\item{tilt}{Angle, in degrees, down from the horizontal plane.}

\item{H}{Height of the observer, in meters.}

\item{R}{Radius of the earth, in kilometers (default is \code{R} at latitude 53 degrees).}
}
\value{
A vector of distances.
}
\description{
Great circle distance based on angle
}
