\name{hadamard}
\alias{hadamard}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Hadamard matrices }
\description{
Returns a Hadamard matrix of dimension larger than the argument.
}
\usage{
hadamard(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{lower bound for size }
}

\value{
A Hadamard matrix
}

\seealso{\code{\link{brrweights}}}
\examples{
image(hadamard(30))
}
\keyword{survey}

