% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_measure.R
\name{image_measure}
\alias{image_measure}
\title{Measure pixels in a set of images}
\usage{
image_measure(
  img_path = "images/",
  log_path = "measures.csv",
  start_index = 1,
  features = c("Other", "Whale (certain)", "Whale (maybe)", "LNG carrier (bow)",
    "LNG carrier (stern)", "Tug (bow)", "Tug (stern)", "Vessel (bow)", "Vessel (stern)"),
  zoom_options = c(30, 50, 60, 70, 80, 90, 100, 125, 150, 175, 200, 300),
  scroll_width = 1700,
  scroll_height = 400,
  filter_to_unmeasured = FALSE
)
}
\arguments{
\item{img_path}{Path to the folder of images. The default assumption is that there is
a folder named "\code{images}" within your working directory.}

\item{log_path}{Path to the \code{csv} file in which to keep pixel measurements.
The default assumption is a file, \code{measures.csv}, within your working directoy.
If this file does not exist in the specified path, the function will create it.}

\item{start_index}{Index of file to start on.}

\item{features}{Character vector of features to choose from. You can measure multiple features on the same image.}

\item{zoom_options}{Numeric vector of zoom options to choose from; the default will be the first one in your vector.}

\item{scroll_width}{Width, in pixels, of image scrolling portion of the app, to customize the app for your specific screen.}

\item{scroll_height}{Height, in pixels, of image scrolling portion of the app, to customize the app for your specific screen.}

\item{filter_to_unmeasured}{Filter images to those without any measurements yet?}
}
\value{
This function will launch a \code{Shiny} app. As you store measurements,
the \code{log_path} file will be updated with those new measurements. Each row of
measurements follows this comma-separated format:
\enumerate{
\item Image filename,
\item Y pixel of horizon/shore
\item Y pixel of whale/boat waterline
\item X pixel of whale/boat (can be useful for refining bearing)
\item System time of when measurement was made
\item Feature (optionally entered manually by user)
\item Latitude (optionally entered manually by user)
\item Longitude (optionally entered manually by user)
\item Elevation (e.g., of drone; optionally entered manually by user)
\item Focal length (optionally entered manually by user)
\item Comment field (optional)
}
}
\description{
This function launches a Shiny app that allows you to quickly measure the pixel location
of the waterline of a whale/boat and the horizon/shoreline directly behind it in a set of images.
}
\details{
This function uses the \code{magick} package to read and display images.

Workflow notes:
\itemize{
\item If your photos are large, you can adjust the zoom options and scroll around
the image using your mousepad.
\item To zoom, click-and-drag to create a rectangle, then single-click within it.
\item To unzoom, single-click anywhere on image.
\item Double-click on the whale/boat waterline to grab its pixel location.
\item Double-click on the horizon/shore line to grab its pixel location.
\item To save these measurements, click the "Save" button that appears.
\item Advance to the next image without saving measurements by clicking the "Skip" button.
\item Close the app by exiting the app window; you may also have to click the 'Stop sign' in
\code{R} or \code{RStudio} to end the command completely.
}
}
