% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_sections.R
\name{img_sections}
\alias{img_sections}
\title{Measure and rate sections in an image}
\usage{
img_sections(
  url_steps,
  url_sections,
  url_cats,
  section_plotter,
  analyst,
  images,
  log_path = "measures.csv",
  scroll_width = 1000,
  scroll_height = 500
)
}
\arguments{
\item{url_steps}{URL to sheet within \code{GoogleSheets} that has step-by-step instructions for each click the user should make on the image.}

\item{url_sections}{URL to sheet within \code{GoogleSheets} that has each section that should be evaluated on the body.}

\item{url_cats}{URL to sheet within \code{GoogleSheets} that has each category by which the section should be evaluated.}

\item{section_plotter}{A function defined by the user, that will plot lines on the image based on the measures clicked on by the user based on \code{url_steps} instructions.}

\item{analyst}{Name of analyst.}

\item{images}{Filepaths to images}

\item{log_path}{Path to where the measurement file should be saved.}

\item{scroll_width}{How wide to make the image scrolling area in the app?}

\item{scroll_height}{How tall?}
}
\value{
A \code{Shiny} app.
}
\description{
Measure and rate sections in an image
}
