% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_to_horizon.R
\name{km_to_horizon}
\alias{km_to_horizon}
\title{Calculate distance to horizon}
\usage{
km_to_horizon(h, R = 6364.5)
}
\arguments{
\item{h}{Height of observer, in meters}

\item{R}{Radius of the earth, in kilometers (default is \code{R} at latitude 53 degrees).}
}
\value{
A list with two elements: \verb{$a} is the angle to the horizon,
in degrees from a horizontal plane,
and \verb{$km} is the distance, in kilometers, from the observer to the horizon.
}
\description{
Calculate distance to horizon
}
