% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htwor.R
\name{mean_with_variance}
\alias{mean_with_variance}
\alias{mean_estimator}
\alias{mean_variance}
\title{Calculate a mean and variance from a two-stage sample}
\usage{
mean_with_variance(
  x,
  psu_index,
  psu_probability,
  psu_size,
  ssu_probability,
  joint_probability
)

mean_estimator(x, psu_index, psu_probability, psu_size, ssu_probability)

mean_variance(
  x,
  mean,
  psu_index,
  psu_probability,
  psu_size,
  ssu_probability,
  joint_probability
)
}
\arguments{
\item{x}{A numeric vector with the variable-of-interest's values.}

\item{psu_index}{A vector with the PSU membership for each element in the \code{x}
vector.}

\item{psu_probability}{A vector with the PSU selection probabilities. It
should be the same length as the number of unique PSUs.}

\item{psu_size}{A vector with the size of each PSU. It should be the same
length as \code{psu_probability}.}

\item{ssu_probability}{A vector with the SSU selection probabilities (ie. the
individual's selection probability given the selection of its PSU).}

\item{joint_probability}{A square matrix of the PSU joint selection
probabilities. It's dimensions should be equal to the number of unique
PSUs. The PSU selection probabilities should be on the diagonal.}

\item{mean}{The survey mean of the variable.}
}
\value{
A named vector with the mean and its variance.
}
\description{
The mean and its variance are calculated following the equations in \href{http://www.ru.ac.bd/wp-content/uploads/sites/25/2019/03/306_06_Sarndal_Model-Assisted-Survey-Sampling-1992.pdf}{Model Assisted Survey Sampling by Sarndal, et al. (1992), pp. 314-5}
}
\section{Functions}{
\itemize{
\item \code{mean_estimator()}: Only calculate the mean.

\item \code{mean_variance()}: Only calculate the variance (given a
pre-calculated mean).

}}
\seealso{
\code{\link[=svymean]{svymean()}} and internal functions in the
\code{survey} package: \code{survey:::ppsvar()}
\code{survey:::htvar.matrix()}, \code{survey:::pi2Dcheck()}.
}
