% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_histogram}
\alias{plot_histogram}
\alias{plot_histogram.probsurvey}
\title{histogram for probabiliy forecasts}
\usage{
plot_histogram(srvObj, year = NULL, quarter = NULL, ...)

\method{plot_histogram}{probsurvey}(srvObj, year = NULL,
  quarter = NULL, id = NULL)
}
\arguments{
\item{srvObj}{a S3-object created by probabilitySPF()}

\item{year}{Year of forecast}

\item{quarter}{quarter of forecast}

\item{id}{forecaster id}
}
\description{
histogram for probabiliy forecasts
}
