% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.r
\name{probabilitySPF}
\alias{probabilitySPF}
\title{download probability forecasts}
\usage{
probabilitySPF(survey = "prgdp")
}
\arguments{
\item{survey}{The survey data to download}
}
\value{
Returns an S3 object of the class probsurvey
}
\description{
download probability forecasts
}
