% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_visibility.R
\name{scan_conditions}
\alias{scan_conditions}
\title{Find average conditions of a scan}
\usage{
scan_conditions(scans, sea, max_vis = 15, verbose = TRUE)
}
\arguments{
\item{scans}{A \code{data.frame} produced by \code{survey_overviews()} or \code{survey_overview()}.}

\item{sea}{A \code{data.frame} produced by \code{survey_overviews()} or \code{survey_overview()}.}

\item{verbose}{Boolean: print progress updates to Console?}
}
\value{
The \code{scans} input with new conditions columns.
}
\description{
Find average conditions of a scan
}
