% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoreline_finder.R
\name{shoreline_finder}
\alias{shoreline_finder}
\title{Find closest shoreline along bearing}
\usage{
shoreline_finder(
  shoreline,
  bearing,
  platform_x = -129.3720833,
  platform_y = 53.2369,
  min_km = 0.5,
  sightline_km = 50,
  verbose = TRUE,
  toplot = FALSE
)
}
\arguments{
\item{shoreline}{An \code{sf} \code{MULTIPOLYGON} of shoreline boundaries to test.}

\item{bearing}{A numeric vector of the bearings of interest, in degrees (0-360).}

\item{platform_x}{Longitude of observation platform, in decimal degrees (the default pertains to Fin Island Research Station).}

\item{platform_y}{Latitude to observation platform. in decimal degrees.}

\item{min_km}{The minimum km to consider;
this allows you to ignore shoreline within a certain distance of your platform.}

\item{sightline_km}{The length of the sightline along your bearing to consider.
Large numbers are safer, but could slow things down if your shoreline is complex.}

\item{verbose}{Boolean: Print updates to Console?}

\item{toplot}{boolean: Create diagnostic plots?}
}
\value{

}
\description{
Find closest shoreline along bearing
}
