% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sighting_distance.R
\name{sighting_distance}
\alias{sighting_distance}
\title{Calculate sighting distance}
\usage{
sighting_distance(
  platform_height,
  km_ref,
  reticles,
  degrees_per_reticle = 0.279,
  R = 6364.5
)
}
\arguments{
\item{platform_height}{Height, in meters, of observation deck (or eye height, if possible) above sea level.}

\item{km_ref}{The distance, in kilometers, to the boundary used as a reference for reticles (horizon or a distant shoreline).}

\item{reticles}{Reticle reading.}

\item{degrees_per_reticle}{The degrees represented by each full reticle.}

\item{R}{Radius of the earth, in kilometers (default is \code{R} at latitude 53 degrees).}
}
\value{
The distance to the sighting, in km.
}
\description{
Calculate sighting distance
}
