% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sighting_distances.R
\name{sighting_distances}
\alias{sighting_distances}
\title{Calculate distances to sightings}
\usage{
sighting_distances(
  sits,
  optics_degrees = list(bigeyes = 0.057296, binocs = 0.279, eyes = 0.279, spotbox =
    0.00060585),
  default_method = "bigeyes",
  R = 6364.5,
  verbose = TRUE
)
}
\arguments{
\item{sits}{A \code{data.frame} of sightings, produced by \code{survey_overview()} or
\code{survey_overviews()}, with an additional column: \code{km_ref} (km to reference boundary).}

\item{optics_degrees}{A named list specifying the degrees-per-reticle conversion
for each optics instrument used in the survey. Spaces and capitalizations will be ignored.}

\item{default_method}{The default optics method to use, in the event that it is missing for
a sighting.}

\item{R}{Radius of the earth, in kilometers (default is \code{R} at latitude 53 degrees).}

\item{verbose}{Boolean: Print updates to Console?}
}
\value{
A vector of distances, in km.
}
\description{
This is a wrapper of \code{sighting_distance()}
}
