% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sighting_locations.R
\name{sighting_locations}
\alias{sighting_locations}
\title{Get sighting locations}
\usage{
sighting_locations(
  bearings,
  kms,
  platform_x = -129.3720833,
  platform_y = 53.2369
)
}
\arguments{
\item{bearings}{Numeric vector of bearings, in degrees.}

\item{kms}{Numeric vector of sightings distances, in km, the same length as \code{bearings}.}

\item{platform_x}{Longitude of observation platform, in decimal degrees (the default pertains to Fin Island Research Station).}

\item{platform_y}{Latitude to observation platform. in decimal degrees.}
}
\value{

}
\description{
Get sighting locations
}
