% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_overview.R
\name{survey_overview}
\alias{survey_overview}
\title{Summarize survey data for a single date}
\usage{
survey_overview(survey_date = NULL)
}
\arguments{
\item{survey_date}{Survey date, in format \code{"yyyy-mm-dd"}. If \code{NULL} the current date
will be used (according to your system's time).}
}
\value{
This function will look for survey data within the \code{data} subfolder
of your working directory and return a list with various summary tables:

\enumerate{
\item \code{scans}: A summary table of scans and their duration.
\item \code{sighting_summary}: Summary metrics for each species observed.
\item \code{sightings}: All data for all sightings, with intuitive column names added.
\item \code{conditions}: All data for all condition zones, with intuitive column names added.
\item \code{comments}: All comments.
\item \code{data}: All data, with two fields added: \code{effort} (indicating whether each row occurs
during systematic effort, with value \code{1}, or off-scan, value \code{0}) and \code{scan_id}
(with a numeric indicator of which scan of the day the row of day correspnds to;
if the row of data occurs outside of a scan, the value with be \code{NA}).
}
}
\description{
Summarize survey data for a single date
}
