% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_overviews.R
\name{survey_overviews}
\alias{survey_overviews}
\title{Summarize survey data for a multiple dates}
\usage{
survey_overviews(
  survey_dates = NULL,
  time_zone = "Canada/Pacific",
  verbose = TRUE
)
}
\arguments{
\item{survey_dates}{A character vector of survey date, in format \code{"yyyy-mm-dd"}.
If not provided or \code{NULL}, the function will process all dates within the \code{data} subfolder of
your working directory.}

\item{time_zone}{The time zone to which you wish to force all date-times in the data
(the data numbers will not be changed at all, just R's interpretation of the time zone using
the \code{lubridate} package).}

\item{verbose}{Boolean: Print updates to Console?}
}
\value{
This function will look for survey data within the \code{data} subfolder
of your working directory and return a list with various summary tables:

\enumerate{
\item \code{scans}: A summary table of scans and their duration.
\item \code{sighting_summary}: Summary metrics for each species observed.
\item \code{sightings}: All data for all sightings, with intuitive column names added.
\item \code{conditions}: All data for all condition zones, with intuitive column names added.
\item \code{comments}: All comments.
\item \code{data}: All data, with two fields added: \code{effort} (indicating whether each row occurs
during systematic effort, with value \code{1}, or off-scan, value \code{0}) and \code{scan_id}
(with a numeric indicator of which scan of the day the row of day correspnds to;
if the row of data occurs outside of a scan, the value with be \code{NA}).
}
}
\description{
A wrapper for \code{survey_overview()}.
}
