% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_sync.R
\name{tide_sync}
\alias{tide_sync}
\title{Find tidal height for survey timestamps}
\usage{
tide_sync(
  df,
  tide,
  date_col = 2,
  tide_col = 3,
  t_offset = 0,
  platform_height = 0,
  verbose = TRUE
)
}
\arguments{
\item{df}{A \code{dataframe} with a \code{date} column (\code{lubridate::as_datetime()}) in the same
time zone as the tidal data below.}

\item{date_col}{The index of the column with datetime data in the tide \code{csv}'s.}

\item{tide_col}{The index of the column with tide height data in the tide \code{csv}'s.}

\item{t_offset}{A time offset, in seconds, to apply to the tide data before synching with \code{df}.}

\item{platform_height}{A height offset, in meters, to apply to the tide data before synching with \code{df}.
This is a shortcut for applying a platform's or observer's height above mean low low tide.}

\item{verbose}{Boolean: Print updates to Console?}

\item{tides}{A \code{data.frame} of tidal date, which must have
a column with date-time information,
format \verb{YYYY-MM-DD hh:mm:ss}, in the same time zone as \code{df},
and a column with tide height information,
in meters above mean low low tide.
For the Gitga'at First Nation area, we retrieve tidal date from the data portal
on the \href{https://www.pac.dfo-mpo.gc.ca/science/charts-cartes/obs-app/observed-eng.aspx?StationID=09130}{Fisheries & Oceans Website},
making sure to set timestamp to local (UTC+8) before saving the \code{csv}'s.
We have supplied this dataset built-in to the package: \code{data(tides)}.}
}
\value{
The same \code{df} that was supplied, now with three new columns:
\code{tide_ht} (the tide height for each row),
\code{tide_diff} (the time difference, in seconds, between the \code{df} timestamp and
the row of tidal data referenced for tide height), and
\code{ht} (the height of the platform above the tidally-adjusted sea level)
}
\description{
Find tidal height for survey timestamps
}
