% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_measure.R
\name{image_measure}
\alias{image_measure}
\title{Measure pixels in a set of images}
\usage{
image_measure(
  img_path = "images/",
  log_path = "measures.csv",
  scroll_width = 1700,
  scroll_height = 400
)
}
\arguments{
\item{img_path}{Path to the folder of images. The default assumption is that there is
a folder named "\code{images}" within your working directory.}

\item{log_path}{Path to the \code{csv} file in which to keep pixel measurements.
The default assumption is a file, \code{measures.csv}, within your working directoy.
If this file does not exist in the specified path, the function will create it.}

\item{scroll_width}{Width, in pixels, of image scrolling portion of the app, to customize the app for your specific screen.}

\item{scroll_height}{Height, in pixels, of image scrolling portion of the app, to customize the app for your specific screen.}
}
\value{
This function will launch a \code{Shiny} app. As you store measurements,
the \code{log_path} file will be updated with those new measurements. Each row of
measurements follows this comma-separated format:
\enumerate{
\item Image filename,
\item Y pixel of horizon/shore
\item Y pixel of whale/boat waterline
\item X pixel of whale/boat (can be useful for refining bearing)
\item System time of when measurement was made
}
}
\description{
This function launches a Shiny app that allows you to quickly measure the pixel location
of the waterline of a whale/boat and the horizon/shoreline directly behind it in a set of images.
}
\details{
This function uses \code{EBImage} to read and display images. It uses \code{exifr} and \code{exiftoolr} to
read image metadata.

Workflow notes:
\itemize{
\item If your photos are large, you can adjust the zoom options and scroll around
the image using your mousepad.
\item Double-click on the whale/boat waterline to grab its pixel location.
\item Single-click on the horizon/shore line to grab its pixel location.
\item To save these measurements, click the "Save" button that appears.
\item The app will automatically advance to the next image.
\item If you can't take both of those measurements for an image, click the "Can't" button to mark
that image as processed as advance to the next image.
\item Advance to the next image without saving measurements by clicking the "Skip" button.
\item Note that by default, the app will only show images that do not yet
have measurements within the \code{log_path} file.
\item Close the app by exiting the app window; you may also have to click the 'Stop sign' in
\code{R} or \code{RStudio} to end the command completely.
}
}
