% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_sections.R
\name{img_sections}
\alias{img_sections}
\title{Measure and rate sections in an image}
\usage{
img_sections(
  url_steps,
  url_sections,
  url_cats,
  section_plotter,
  analyst,
  images,
  log_path = "measures.csv",
  scroll_width = 1000,
  scroll_height = 500
)
}
\arguments{
\item{url_steps}{desc}

\item{url_sections}{desc}

\item{url_cats}{desc}

\item{section_plotter}{desc}

\item{analyst}{desc}

\item{images}{desc}

\item{log_path}{desc}

\item{scroll_width}{desc}

\item{scroll_height}{desc}
}
\value{
desc
}
\description{
Measure and rate sections in an image
}
