% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.r
\name{individualSPF}
\alias{individualSPF}
\title{Download survey data}
\usage{
individualSPF(survey = "NGDP", variable = "NGDP1")
}
\arguments{
\item{survey}{The survey you wish to download}

\item{variable}{The variable name indicating the forecast horizon, e.g. "NGDP1" or "CPI2".}
}
\value{
individual SPF returns an object of class "survey".

An object of class "survey" is a list containing the following components
\item{survey}{The name of the survey}
\item{type}{Type of the survey, i.e. mean, median, growth, or individual (see individualSPF)}
\item{variable}{Name of the variable indicating the forecast horizon, e.g. "NGDP1", or "NGDP2", downloaded. If "all", all forecast horizons were downloaded}
\item{sseries}{An object of the class ts containing the downloaded series}
}
\description{
Download survey data
}
\examples{
tmp <- downloadSPF(survey = "CPI", type = "growth")
}
